requirements.txt
alembic==1.0.11
celery==4.4.6
factory-boy==2.12.0
flake8==3.8.3
Flask==1.1.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
marshmallow-sqlalchemy==0.23.1
pytest-cov==2.10.0
pytest-factoryboy==2.0.3
pytest-flask-sqlalchemy==1.0.2
tox==3.16.0
virtualenv==20.0.25
webargs==5.3.2